pyproject.toml
[tool.poetry]
name = "zeroapi"
version = "0.1.0"
description = ""
authors = ["YAndrii27 <AndriiYashchishen@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.20"
uvicorn = "^0.23.2"
envparse = "^0.2.0"
asyncpg = "^0.28.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
quart = "^0.18.4"
quart-schema = "^0.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"